+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Fix BoxedTabs style
 - Update some color usage
   - SONAR-12539 Update leak color of timeline
+  - SONAR-12541 Rewrite GlobalMessages to @emotion and clean colors and css
 
 ## 0.0.31